+++ CHANGELOG.md
@@ -7,9 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.0.3] - 2022-07-16
 ### Added
-- added bubble sort
-- added insertion sort
-- added bubble sort
+- quick sort
+- bubble sort
+- insertion sort
+- bubble sort
 - functionality to sorting alogrithms view
 
 ## [0.0.2] - 2022-07-16